This commit was manufactured by cvs2git to create tag 'old_build_system'.

+++ src/IscDbc/AsciiBlob.cpp
@@ -77,7 +77,7 @@ int AsciiBlob::release()
 	return useCount;
 }
 
-int AsciiBlob::length()
+uint64 AsciiBlob::length()
 {
 	if (!populated)
 		populate();
